Update
add JoyStick
add speed limit
add drag
add rotation drag
add gears
fix rotation to scale with car speed

+++ TestBox/Assets/Scripts/collider.cs
@@ -13,7 +13,7 @@ public class collider : MonoBehaviour
     public Text yellowText;
     public Text greenText;
 
-    public int waitTime;
+    public float waitTime;
     private Boolean safeToPass;
 
     private void Start()
